+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changes
 * [Developer]: Cleaned up old use deprecated method `CrudService.update(Long, Map)` and removed the method from the project.
 * [API]: Added configurable maximum number of workflows IRIDA will schedule at a time `irida.workflow.max-running`.  Default 4.
 * [UI]: Fixed login error message that requested an email instead of username.
+* [Developer]: Added quick failure on detection of an upload error in Galaxy.
+* [Developer]: Added ability to adjust size of thread pool for polling Galaxy after uploading files with `galaxy.library.upload.threads`.
 
 0.13.0 to 0.14.0
 ----------------